documents/Esame_SAC/requirements.txt
flask
flask_restful
google-cloud-firestore
google-cloud-pubsub
gunicorn
requests